# collections: speed up display of tags for tag filter