@TurboGit 2021-05-21
Nice! Let's try to have this in 3.6. I'll review and asked @phweyland to review also as pretty good in SQL.