@phweyland 2021-05-21
Nice indeed, making the string comparison only on already filtered/reduced list should speed up things.
I've compared the queries and haven't found any discrepancy.
If you let me some time, I'll make some tests later today.