@ralfbrown 2021-05-21
I hadn't been aware of `memory.darktable_tags`.  Using it does indeed save a bit of time (now down to 0.63-0.64 seconds).  When I looked at the code manipulating it, I was worried that the lazy init/update would cause an issue, but apparently there's always an invocation of a function that calls the init before the treeview is built.  OTOH, replacing the other occurrence of the substring check that I found with `memory.darktable_tags` resulted in an incorrect thumbtable for the "tag: untagged" filter until I switched to darkroom and back out to lighttable, so I'm leaving that one unchanged.