@phweyland 2021-05-22
yes, I was starting with dt when I implemented this. The initialization should be made at the start of darktable.  I'll try to fix it.